examples: moved; get_env_var, get_user_config, channels_to_join, handle_message -> include/mod
