+++ examples/async_std_demo.rs
@@ -1,40 +1,13 @@
 // NOTE: this demo requires `--features="async-std async-std/attributes"`.
+mod include;
 use anyhow::Context;
 
+use crate::include::{channels_to_join, get_user_config, handle_message};
 use twitchchat::{
     commands, connector, messages,
     runner::{AsyncRunner, Status},
     UserConfig,
 };
-
-fn get_env_var(key: &str) -> anyhow::Result<String> {
-    std::env::var(key).with_context(|| format!("please set `{}`", key))
-}
-
-fn get_user_config() -> anyhow::Result<twitchchat::UserConfig> {
-    let name = get_env_var("TWITCH_NAME")?;
-    let token = get_env_var("TWITCH_TOKEN")?;
-
-    // you need a `UserConfig` to connect to Twitch
-    let config = UserConfig::builder()
-        // the name of the associated twitch account
-        .name(name)
-        // and the provided OAuth token
-        .token(token)
-        // and enable all of the advanced message signaling from Twitch
-        .enable_all_capabilities()
-        .build()?;
-
-    Ok(config)
-}
-
-fn channels_to_join() -> anyhow::Result<Vec<String>> {
-    let channels = get_env_var("TWITCH_CHANNEL")?
-        .split(',')
-        .map(ToString::to_string)
-        .collect();
-    Ok(channels)
-}
 
 async fn connect(user_config: &UserConfig, channels: &[String]) -> anyhow::Result<AsyncRunner> {
     // create a connector using ``async_std``, this connects to Twitch.
@@ -60,43 +33,6 @@ async fn connect(user_config: &UserConfig, channels: &[String]) -> anyhow::Resul
     }
 
     Ok(runner)
-}
-
-async fn handle_message(msg: messages::Commands<'_>) {
-    use messages::Commands::*;
-    // All sorts of messages
-    match msg {
-        // This is the one users send to channels
-        Privmsg(msg) => println!("[{}] {}: {}", msg.channel(), msg.name(), msg.data()),
-
-        // This one is special, if twitch adds any new message
-        // types, this will catch it until future releases of
-        // this crate add them.
-        Raw(_) => {}
-
-        /// These happen when you initially connect
-        IrcReady(_) => {}
-        Ready(_) => {}
-        Cap(_) => {}
-
-        // and a bunch of other messages you may be interested in
-        ClearChat(_) => {}
-        ClearMsg(_) => {}
-        GlobalUserState(_) => {}
-        HostTarget(_) => {}
-        Join(_) => {}
-        Notice(_) => {}
-        Part(_) => {}
-        Ping(_) => {}
-        Pong(_) => {}
-        Reconnect(_) => {}
-        RoomState(_) => {}
-        UserNotice(_) => {}
-        UserState(_) => {}
-        Whisper(_) => {}
-
-        _ => {}
-    }
 }
 
 async fn main_loop(mut runner: AsyncRunner) -> anyhow::Result<()> {
