+++ examples/simple_bot.rs
@@ -1,8 +1,10 @@
 // note this uses `smol`. you can use `tokio` or `async_std` or `async_io` if you prefer.
+mod include;
 use anyhow::Context as _;
 use std::collections::HashMap;
 
 // extensions to the Privmsg type
+use crate::include::{channels_to_join, get_user_config};
 use twitchchat::PrivmsgExt as _;
 use twitchchat::{
     messages::{Commands, Privmsg},
@@ -140,35 +142,4 @@ impl Bot {
         }
         input.splitn(2, ' ').next()
     }
-}
-
-// some helpers for the demo
-fn get_env_var(key: &str) -> anyhow::Result<String> {
-    std::env::var(key).with_context(|| format!("please set `{}`", key))
-}
-
-// channels can be either in the form of '#museun' or 'museun'. the crate will internally add the missing #
-fn channels_to_join() -> anyhow::Result<Vec<String>> {
-    let channels = get_env_var("TWITCH_CHANNEL")?
-        .split(',')
-        .map(ToString::to_string)
-        .collect();
-    Ok(channels)
-}
-
-fn get_user_config() -> anyhow::Result<twitchchat::UserConfig> {
-    let name = get_env_var("TWITCH_NAME")?;
-    let token = get_env_var("TWITCH_TOKEN")?;
-
-    // you need a `UserConfig` to connect to Twitch
-    let config = UserConfig::builder()
-        // the name of the associated twitch account
-        .name(name)
-        // and the provided OAuth token
-        .token(token)
-        // and enable all of the advanced message signaling from Twitch
-        .enable_all_capabilities()
-        .build()?;
-
-    Ok(config)
 }
